UPDATE web/apis
* Small tweaks to the way that the table schema is displayed
* It is now possible to update a variety of columns from the admin page, however there is only a minimum of safeguarding, so it is definetly possible to mess something up, however it is not possible to set the password, or a few other things.

Considering that this admin table is fully working, and it is possible to set how much a user is allowed to store, it might be about time to fix the bugs in the uploader, and enable signups (which will be a button in the admin interface/command flag [admin will override]).

Plus could add a sort of overview of the health of the service, with an uptime counter, and then also if there are any updates that are overdew.