BUG FIX
Now when playback ends, using the existing function 'PLAYBACK_stop_playback', it will now do something and clears the top, and clears the queue, and puts up a sort of blocking card so that the unused LCD doesn't look ugly.

This is something that I should have had a lot earlier, never the less it is here now.