STYLING
Polished up a lot of rough edges, and ensured that everything to do with the settings is mostly fine as of the moment. In the future of course it would be good to have the ability to make the nav menu bigger or smaller, so that it doesn't always take up so much of the space, but that's all for the future.