UPDATE search
Have now almost completed search. The backend is complete, the frontend is almost complete, the two functions which either take you to a project or a song are both fully functional and feature complete, with their respective very nice stylings.

Just need to capture the tabs, and move those correctly to make it accessible, and meaningful to people who like using the keyboard as well.