UPDATE search
Search is now complete, it is fully functional, you can tab into search from the respective areas. It's fast, useful, and has a little bit of flair. Which means that that is another item ticked off of the list, however a new item has been added :(, which is however essential to a good experience.