UPDATE projectView
Have been working on making the process of moving items around in a project more smooth (not having to rely on the returned JSON to rebuild the array, to make it accurate). And when dragging songs upwards this has been fixed. However when moving them downwards, something incorrect is happening, so therefore the system that I have devised/adapted that goes through the current Details and then compares with the server accurate one and then if a difference is detected updates that locally. This is currently being fleshed out, since there are a few pieces of iffy code surrounding this principle in projectview, so this is something that I will work on to now make it more streamlined, and increasing code reuse.