UPDATE projectVIEW
Have now completed the project of refactoring some of projectView
Have also now implemented a sort of local cache for projects, this means that it will load the last copy of Details it has, and if it equals what is on the server, no further action is needed. It's really rather quiet good.