UPDATE musicObjectGrid
Have now implemented a system similar to the one in projectView, that allows me to cache the first 30 items in the library, and then have it retrieve it when a user returns to that page, and then just update the details that are different when the data gets checked from the server.
It works pretty well, and definetly speeds up the experience of back navigating from a project, to the MOG on mobile. And since this 'costs' nothing, it's a pretty good option. There might be cases where slow network speeds cause a bad perception of the things being updated after a half second or whatever it is, but that is something that can be investigated, and the remedies for such a condition.