UPDATE/CREATE musicEnjoy
I have now started working on the musicEnjoy mode which would basically be a focused view, in which just the image, basic player controls and not much more would be visible. And hopefully also with some sort of visualiser/reactor to the music and based on the colours available from the playback image.

I now have the buttons that open and close this mode working, and that's all very simple, and now I am working on implementing a basic, rough layout of what it could look like.