UPDATE musicEnjoy
The music visualiser is now functional and it is so very cool, and on good browsers like chrome, it doesn't result in any performance issues. I tried to getting it working in such a way that my visualiser code should be reusable for multiple destinations, so now I want to see if it will work for the mobile version, since that is something that would be very cool indeed if that just worked.