UPDATE musicEnjoy
Have now fixed the bug that wouldn't serve the mask.png image, which resulted in the masked image for mobile musicEnjoy being hidden.
Have also discovered that the visualiser on mobile has infact been working this whole time, it has just been slightly more subtle/smooth? So I might work on making it more distinctive, but it is actually working.