UPDATE frontend
Have now added funtional menu items to the song menu button in the LCD.
Have now added a return promise option to 'MENUdisplay()', which means that it is now possible to get which condition was clicked returned to the area of code that originaly called 'MENUdisplay()', rather than having to add an export function and patching it into menu.js every time a new menu function was required.
Have now removed the menu.js code that showed the old mobile 'musicEnjoy' mode and the project is now lighter as a result.