UPDATE routing
Have now updated the docker-compose and such so that web-content is stored on a local driver volume, which means that from this point on it will be very easy to live update the web-content during development.
Have now updated the router so that of go_api so that it will serve all paths, and this is just how it should have been from the start.